pyproject.toml
[tool.poetry]
name = "bad-apple"
version = "0.1.0"
description = ""
authors = ["CharmfulProgrammer"]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.6.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
